# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A social media API service built with Go featuring user management, posts, comments, and following functionality",
        "title": "Gopher Social API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
...
        "/posts": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new post with title, content and optional tags. Requires JWT authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Create a new post",
                "parameters": [
                    {
                        "description": "Post creation data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_payload.CreatePOstPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created post",
                        "schema": {
                            "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_posts.Post"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/posts/feed": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a paginated feed of posts from followed users and own posts. Requires JWT authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feed"
                ],
                "summary": "Get user's post feed",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "example": 10,
                        "description": "Number of posts per page (1-20)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "example": 0,
                        "description": "Number of posts to skip",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort order (asc/desc)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"golang\"",
                        "description": "Search in title and content",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"golang,programming\"",
                        "description": "Comma-separated list of tags",
                        "name": "tags",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"2024-01-01 00:00:00\"",
                        "description": "Posts created after this date",
                        "name": "since",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\"2024-12-31 23:59:59\"",
                        "description": "Posts created before this date",
                        "name": "until",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User's post feed",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_posts.Feed"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/posts/{postID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get detailed information about a specific post including comments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Get post by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Post ID",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post information with comments",
                        "schema": {
                            "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_posts.Post"
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a specific post by its ID. Requires JWT authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Delete a post",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Post ID",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Post successfully deleted"
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update the title and/or content of an existing post. Requires JWT authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "Update a post",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Post ID",
                        "name": "postID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Post update data",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_payload.UpdatePostPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated post",
                        "schema": {
                            "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_posts.Post"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - invalid or missing token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_orangeMangoDimz_go-social_internal_entities_comments.Comment": {
            "description": "Comment on a social media post",
            "type": "object",
            "properties": {
                "content": {
                    "description": "Comment content",
                    "type": "string",
                    "example": "Great post!"
                },
                "created_at": {
                    "description": "Comment creation timestamp",
                    "type": "string",
                    "example": "2024-01-01 12:00:00"
                },
                "id": {
                    "description": "Comment ID",
                    "type": "integer",
                    "example": 1
                },
                "post_id": {
                    "description": "ID of the post this comment belongs to",
                    "type": "integer",
                    "example": 123
                },
                "user": {
                    "description": "User who made the comment",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_users.User"
                        }
                    ]
                },
                "user_id": {
                    "description": "ID of the user who made the comment",
                    "type": "integer",
                    "example": 456
                }
            }
        },
        "github_com_orangeMangoDimz_go-social_internal_entities_payload.CreatePOstPayload": {
            "description": "Request payload for creating a new post",
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "content": {
                    "description": "Post content (max 1000 characters)",
                    "type": "string",
                    "maxLength": 1000,
                    "example": "This is the content of my post"
                },
                "tags": {
                    "description": "Post tags",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "golang",
                        "programming"
                    ]
                },
                "title": {
                    "description": "Post title (max 100 characters)",
                    "type": "string",
                    "maxLength": 100,
                    "example": "My First Post"
                }
            }
        },
...
        "github_com_orangeMangoDimz_go-social_internal_entities_payload.UpdatePostPayload": {
            "description": "Request payload for updating an existing post",
            "type": "object",
            "properties": {
                "content": {
                    "description": "Updated post content (max 1000 characters)",
                    "type": "string",
                    "maxLength": 1000,
                    "example": "Updated post content"
                },
                "title": {
                    "description": "Updated post title (max 100 characters)",
                    "type": "string",
                    "maxLength": 100,
                    "example": "Updated Post Title"
                }
            }
        },
...
        "github_com_orangeMangoDimz_go-social_internal_entities_posts.Feed": {
            "description": "Post feed item with comment count",
            "type": "object",
            "properties": {
                "comments": {
                    "description": "Comments on this post",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_comments.Comment"
                    }
                },
                "content": {
                    "description": "Post content",
                    "type": "string",
                    "example": "This is my post content"
                },
                "created_at": {
                    "description": "Post creation timestamp",
                    "type": "string",
                    "example": "2024-01-01 12:00:00"
                },
                "id": {
                    "description": "Post ID",
                    "type": "integer",
                    "example": 1
                },
                "tags": {
                    "description": "Post tags",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "golang",
                        "programming"
                    ]
                },
                "title": {
                    "description": "Post title",
                    "type": "string",
                    "example": "My First Post"
                },
                "total_comment": {
                    "description": "Total number of comments on this post",
                    "type": "integer",
                    "example": 5
                },
                "updated_at": {
                    "description": "Post last update timestamp",
                    "type": "string",
                    "example": "2024-01-01 12:30:00"
                },
                "user": {
                    "description": "User who created the post",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_users.User"
                        }
                    ]
                },
                "user_id": {
                    "description": "ID of the user who created the post",
                    "type": "integer",
                    "example": 123
                },
                "version": {
                    "description": "Post version for optimistic locking",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "github_com_orangeMangoDimz_go-social_internal_entities_posts.Post": {
            "description": "Social media post with content, tags and metadata",
            "type": "object",
            "properties": {
                "comments": {
                    "description": "Comments on this post",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_comments.Comment"
                    }
                },
                "content": {
                    "description": "Post content",
                    "type": "string",
                    "example": "This is my post content"
                },
                "created_at": {
                    "description": "Post creation timestamp",
                    "type": "string",
                    "example": "2024-01-01 12:00:00"
                },
                "id": {
                    "description": "Post ID",
                    "type": "integer",
                    "example": 1
                },
                "tags": {
                    "description": "Post tags",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "golang",
                        "programming"
                    ]
                },
                "title": {
                    "description": "Post title",
                    "type": "string",
                    "example": "My First Post"
                },
                "updated_at": {
                    "description": "Post last update timestamp",
                    "type": "string",
                    "example": "2024-01-01 12:30:00"
                },
                "user": {
                    "description": "User who created the post",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_users.User"
                        }
                    ]
                },
                "user_id": {
                    "description": "ID of the user who created the post",
                    "type": "integer",
                    "example": 123
                },
                "version": {
                    "description": "Post version for optimistic locking",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "github_com_orangeMangoDimz_go-social_internal_entities_users.Role": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_orangeMangoDimz_go-social_internal_entities_users.User": {
            "description": "User account information",
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Account creation timestamp",
                    "type": "string",
                    "example": "2024-01-01 12:00:00"
                },
                "email": {
                    "description": "Email address",
                    "type": "string",
                    "example": "johndoe@example.com"
                },
                "id": {
                    "description": "User ID",
                    "type": "integer",
                    "example": 1
                },
                "is_active": {
                    "type": "boolean"
                },
                "role": {
                    "$ref": "#/definitions/github_com_orangeMangoDimz_go-social_internal_entities_users.Role"
                },
                "role_id": {
                    "type": "integer"
                },
                "username": {
                    "description": "Username",
                    "type": "string",
                    "example": "johndoe"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key for authorization",
            "type": "apiKey",
            "name": "X-API-KEY",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Enter your bearer token in the format **Bearer \u0026lt;token\u0026gt;**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}